// 03-platform/src/schemas/crm.d.ts
export declare const CrmContactSchema: z.ZodObject<{
    id: z.ZodString;
    type: z.ZodDefault<z.ZodEnum<["investor", "producer", "partner", "vendor", "legal", "internal", "other"]>>;
    email: z.ZodString;
    first_name: z.ZodString;
    last_name: z.ZodString;
    title: z.ZodOptional<z.ZodString>;
    company_id: z.ZodOptional<z.ZodString>;
    tags: z.ZodDefault<z.ZodArray<z.ZodString, "many">>;
    phone: z.ZodOptional<z.ZodString>;
    whatsapp: z.ZodOptional<z.ZodString>;
    telegram: z.ZodOptional<z.ZodString>;
    source: z.ZodDefault<z.ZodEnum<["referral", "conference", "outbound", "inbound", "partner", "whatsapp", "other"]>>;
    preferred_channel: z.ZodDefault<z.ZodEnum<["whatsapp", "email", "platform", "call"]>>;
    status: z.ZodDefault<z.ZodEnum<["active", "inactive", "do-not-contact", "unsubscribed"]>>;
    created_at: z.ZodString;
    last_contact_at: z.ZodOptional<z.ZodString>;
    notes: z.ZodOptional<z.ZodString>;
}, "strip", z.ZodTypeAny, {
    email: string;
    type: "legal" | "other" | "investor" | "producer" | "partner" | "vendor" | "internal";
    status: "active" | "inactive" | "do-not-contact" | "unsubscribed";
    id: string;
    first_name: string;
    last_name: string;
    tags: string[];
    source: "other" | "whatsapp" | "partner" | "outbound" | "inbound" | "referral" | "conference";
    preferred_channel: "email" | "call" | "whatsapp" | "platform";
    created_at: string;
    title?: string | undefined;
    whatsapp?: string | undefined;
    notes?: string | undefined;
    company_id?: string | undefined;
    phone?: string | undefined;
    telegram?: string | undefined;
    last_contact_at?: string | undefined;
}, {
    email: string;
    id: string;
    first_name: string;
    last_name: string;
    created_at: string;
    type?: "legal" | "other" | "investor" | "producer" | "partner" | "vendor" | "internal" | undefined;
    status?: "active" | "inactive" | "do-not-contact" | "unsubscribed" | undefined;
    title?: string | undefined;
    whatsapp?: string | undefined;
    notes?: string | undefined;
    company_id?: string | undefined;
    tags?: string[] | undefined;
    phone?: string | undefined;
    telegram?: string | undefined;
    source?: "other" | "whatsapp" | "partner" | "outbound" | "inbound" | "referral" | "conference" | undefined;
    preferred_channel?: "email" | "call" | "whatsapp" | "platform" | undefined;
    last_contact_at?: string | undefined;
}>;
// ...
export type CrmContact = z.infer<typeof CrmContactSchema>;